+++ scripts/mods/BetterInventory/BetterInventory_data.lua
@@ -1,4 +1,4 @@
-local MOD_VERSION = "3.3.1"
+local MOD_VERSION = "3.4.0"
 local mod = get_mod("BetterInventory")
 local DEFAULT_OPERATIVE_SLOT_CAPACITY = 10
 local MAX_REASONABLE_OPERATIVE_SLOT_CAPACITY = 64
@@ -1381,6 +1381,18 @@ return {
 						setting_id = "melk_views_group",
 						type = "group",
 						sub_widgets = {
+							{
+								setting_id = "enable_melk_limited_grid",
+								tooltip = "enable_melk_limited_grid_tooltip",
+								type = "checkbox",
+								default_value = true,
+							},
+							{
+								setting_id = "enable_melk_multi_operative_grid",
+								tooltip = "enable_melk_multi_operative_grid_tooltip",
+								type = "checkbox",
+								default_value = true,
+							},
 							{
 								setting_id = "melk_auto_favorite_purchased_items",
 								tooltip = "melk_auto_favorite_purchased_items_tooltip",